+++ middlewares/securityKeyMiddleware.js
@@ -2,7 +2,7 @@ const { NotAuthorizedError } = require("../errors/notAuthorizedError");
 
 module.exports = (req, res, next) => {
     if (req.headers["khokha-security-key"] !== process.env.KHOKHA_SECURITY_KEY) {
-        next(new NotAuthorizedError("Unauthorized Request"));
+        return next(new NotAuthorizedError("Unauthorized Request"));
     }
     next();
-}
+};
